ability to open files directly from mpq archive
git-svn-id: http://localhost:88/svn/uwow/mpqnav/trunk@559 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ MPQNav/Game1.cs
@@ -337,7 +337,7 @@ namespace MPQNav {
 			view = Matrix.CreateLookAt(cameraPosition, avatarPosition, new Vector3(0.0f, 1.0f, 0.0f));
 
 			Viewport viewport = graphics.GraphicsDevice.Viewport;
-			float aspectRatio = (float)viewport.Width / (float)viewport.Height;
+			float aspectRatio = viewport.Width / (float)viewport.Height;
 
 			proj = Matrix.CreatePerspectiveFieldOfView(viewAngle, aspectRatio, nearClip, farClip);
 		}
